fix style

+++ modules/task_1/vedrukov_p_freq_symbol_in_a_str/freq_symb.cpp
@@ -4,20 +4,20 @@
 #include <string>
 #include "../../../modules/task_1/vedrukov_p_freq_symbol_in_a_str/freq_symb.h"
 
-int getFreq(std::string str, char symb) {
+double getFreq(std::string str, char symb) {
     int Proc_num;
     int Proc_rank;
     MPI_Comm_size(MPI_COMM_WORLD, &Proc_num);
     MPI_Comm_rank(MPI_COMM_WORLD, &Proc_rank);
     if (Proc_num <= 1) {
         int num = 0;
-        for (int i = 0; i < str.length(); i++) {
+        for (int i = 0; i < static_cast<int>str.length(); i++) {
             if (str[i] == symb)
                 num++;
         }
         return num / str.length();
     }
-    int substrlen = str.length() / (Proc_num - 2);
+    int substrlen = static_cast<int>str.length() / (Proc_num - 2);
     int last_substrlen = str.length() % (Proc_num - 2);
     int num_symb = 0;
     int all_num_sumb = 0;
